+++ API/MLUpdates.lua
@@ -45,14 +45,19 @@ end
 function MLModule:ShouldAutoAward (item, quality)
    if not item then return false end
    local itemid = addon:GetItemIDFromLink(item)
-   
+   local db = addon:Getdb()
+
    -- Check always list first
-   if itemid and Classic.Lists.AlwaysList[itemid] then
-      return true, "normal", db.autoAwardTo
+   if itemid and db.autoAward and db.alwaysAutoAwardItems[itemid] then
+      if UnitInRaid(db.autoAwardTo) or UnitInParty(db.autoAwardTo) then -- TEST perhaps use self.group?
+         return true, "normal", db.autoAwardTo
+      else
+         addon:Print(L["Cannot autoaward:"])
+         addon:Print(format(L["Could not find 'player' in the group."], db.autoAwardTo))
+      end
    end
 
    -- Check for rep item auto award
-   local db = addon:Getdb()
    if not db.autoAwardRepItems then
       return orig_ShouldAutoAward(MLModule, item, quality)
    end